Move dataclasses to models subfolder
