+++ tdameritrade/orders/__init__.py
@@ -1,4 +1,4 @@
-from . import orders
+from . import order_builder
 
 
-__all__ = ["orders"]
+__all__ = ["order_builder"]
