+++ tdameritrade/orders/models/activities.py
@@ -2,7 +2,7 @@ from dataclasses import dataclass
 from typing import List
 
 from .base import BaseOrder
-from .constants import ActivityType, ExecutionType
+from ..constants import ActivityType, ExecutionType
 
 
 @dataclass
@@ -12,7 +12,7 @@ class OrderActivity(BaseOrder):
 
 @dataclass(frozen=True)
 class ExecutionLeg(BaseOrder):
-    """ExecutionLeg using in Execution
+    """ExecutionLeg used in Execution
     """
 
     legId: int = None
